feat: update tour migration

+++ database/migrations/2024_12_12_093208_create_tour_packages_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignUuid('category_id')->constrained()->onDelete('cascade');
             $table->string('name');
             $table->text('description');
-            $table->decimal('price', 15, 2);
+            $table->integer('price');
             $table->integer('duration_days');
             $table->integer('duration_hours');
             $table->string('location');
